Development

+++ src/slices/stayThunks/deleteStayThunk.ts
@@ -1,27 +1,16 @@
-import {
-    collection,
-    doc,
-    setDoc,
-    updateDoc,
-    getDocs,
-    getFirestore,
-    writeBatch
-} from "@firebase/firestore";
-import { auth, firestore, storage } from "@/lib/firebase";
-import { createFile } from "@/lib/utils";
-import { getDoc } from "firebase/firestore";
+import {doc, writeBatch} from "@firebase/firestore";
+import {firestore} from "@/lib/firebase";
 import {createAsyncThunk} from "@reduxjs/toolkit";
-import {deleteImage, getCurrentUser, uploadImage} from "@/data/hotelsData";
-import {Stay} from "@/lib/types";
+import {deleteImage, getCurrentUser} from "@/data/hotelsData";
+import {Hotel, Stay} from "@/lib/types";
 
 export const deleteStayAsync = createAsyncThunk(
     'stay/deleteStay',
     async (stay: Stay, { rejectWithValue }) => {
         try {
             const user = getCurrentUser();
-            const firestore = getFirestore();
             const originStayRef = doc(firestore, 'hosts', user.uid, 'stays', stay.id);
-            const userDocRef = doc(firestore, 'hosts', user.uid);
+            const publicStaysRef = doc(firestore, 'stays', stay.id);
 
             if (stay.published) {
                 throw new Error('You have to unpublish the stay before deleting it.');
@@ -29,13 +18,14 @@ export const deleteStayAsync = createAsyncThunk(
 
             const batch = writeBatch(firestore);
 
-            // Delete room images and documents
-            for (const room of stay.rooms) {
-                await deleteImage(`${stay.id}/${room.id}/poster`);
-                await Promise.all(room.images.map((_:string, index: number) => deleteImage(`${stay.id}/${room.id}/image-${index}`)));
+            if (stay.type === 'Hotel'){  // Delete room images and documents
+                for (const room of (stay as Hotel).rooms) {
+                    await deleteImage(`${stay.id}/${room.id}/poster`);
+                    await Promise.all(room.images.map((_: string, index: number) => deleteImage(`${stay.id}/${room.id}/image-${index}`)));
 
-                const roomRef = doc(originStayRef, 'rooms', room.id);
-                batch.delete(roomRef);
+                    const roomRef = doc(originStayRef, 'rooms', room.id);
+                    batch.delete(roomRef);
+                }
             }
 
             // Delete stay images
@@ -46,6 +36,7 @@ export const deleteStayAsync = createAsyncThunk(
 
             // Delete stay document
             batch.delete(originStayRef);
+            batch.delete(publicStaysRef)
             await batch.commit();
 
             return stay.id;
